+++ client_impl.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package driver
@@ -210,4 +208,28 @@ func (c *client) SetLogLevels(ctx context.Context, logLevels LogLevels, opts *Lo
 	}
 
 	return nil
+}
+
+// GetLicense returns license of an ArangoDB deployment.
+func (c *client) GetLicense(ctx context.Context) (License, error) {
+	result := License{}
+	req, err := c.conn.NewRequest(http.MethodGet, "_admin/license")
+	if err != nil {
+		return result, WithStack(err)
+	}
+
+	applyContextSettings(ctx, req)
+	resp, err := c.conn.Do(ctx, req)
+	if err != nil {
+		return result, WithStack(err)
+	}
+	if err := resp.CheckStatus(http.StatusOK); err != nil {
+		return result, WithStack(err)
+	}
+
+	if err := resp.ParseBody("", &result); err != nil {
+		return result, WithStack(err)
+	}
+
+	return result, nil
 }